+++ packages/react-dom/src/ReactDOMFloat.js
